Allow selection of auth manager in run.sh script
This is probably only useful for testing, since this script does not
really allow for any way to configure the Basic auth manager's user
credentials file nor any useful way to connect to an OpenShift
authentication server. But, it's still helpful to be able to check that
the container application and the -web client properly detect the
configured auth manager and react accordingly.