Set up "connected" commands to establish their own target connections
Add methods to base classes AbstractConnectedCommand and
AbstractRecordingCommand for establishing new connections to target JVMs
and performing actions using those connections. This allows the commands
to be executed without first requiring state to track the
"active"/current target connection. The ListCommand is modified to
become stateless this way as a first test.