Use synchronous/blocking handling for saved reports
Use synchronous filesystem calls for creating cached reports to ensure
that rapid duplicate requests for the same report do not potentially
receive a partially-written cached file. This will somewhat increase
latency (due to IO to write the report to the filesystem) on saved recording
report requests when the cache is missed, but ensures cache consistency and has
no impact on requests where the cache is hit.

Also, if duplicate requests are made while the report is still being
generated or written to cache, then the subsequent requests will cause
new report generations to occur since there is no cached file to
retrieve and the cache will be missed. This is also a consequence of
enforcing cache consistency while allowing the Vert.x server to continue
to serve requests while the first request is being processed.