Make small change to CommandRegistryImpl validate()
Previously, if the command was found to not be registered, the thrown
exception would not contain an `errorMessage` (i.e. it would use
the empty constructor). I did this because the method used to check
for a command's registration (`isCommandAvailable()`) already outputs
an error message when returning false.

But rethinking it, it probably makes more sense to have the validate
method to just always include an error message, even if it is a bit
redundant. We could also change AbstractCommandExecutor to work more
like WsCommandExecutor by having it check for registration and
availability of commands before validating them, though that is
probably a lot of work for not much gain.