Change Response Message from `validate()` (#206)
* Create FailedValidationException class

* Change validate methods to be void and throw FailedValidationException

* Make changes to validate methods in registry implementations

* Change command executors to work with changes to validate

* Make small change to CommandRegistryImpl validate()

Previously, if the command was found to not be registered, the thrown
exception would not contain an `errorMessage` (i.e. it would use
the empty constructor). I did this because the method used to check
for a command's registration (`isCommandAvailable()`) already outputs
an error message when returning false.

But rethinking it, it probably makes more sense to have the validate
method to just always include an error message, even if it is a bit
redundant. We could also change AbstractCommandExecutor to work more
like WsCommandExecutor by having it check for registration and
availability of commands before validating them, though that is
probably a lot of work for not much gain.

* Modify unit tests for executors to work with new validate

* Modify unit tests for registry implementations

* In Command.validate(), add back in ClientWriter printing error messages

* Change tests to reflect the re-added cw.println() in validate()

* Make naming for validate argc tests more consistent

* Small fix to COMMANDS.md

The `wait-for` command was missing a `targetId` parameter.

Changed the `target` parameter in the `list` command to be `targetId`,
to be more consistent.

* Change example commands in BatchModeExecutorTest to existing commands

* Make small final fixes

* Fix typo in COMMANDS.md

* Use implicit return in empty `validate()` methods

* Fix short-circuiting behaviour in `validate()`

* Remove print statement from `validateEvents()`

* Use StringJoiner instead of manual delimiters