Replace JMX_AUTH and DISABLE_HTTPS with FORCE_INSECURE
CONTAINER_JFR_FORCE_INSECURE environment variable disables SSL
certificate generation, which in turn disables JMX SSL as well as
WebServer HTTPS/WSS, as well as JMX auth credential generation, if the
variable is set to "true". Otherwise, or if CONTAINER_JFR_RJMX_USER or
CONTAINER_JFR_RJMX_PASS are set to non-empty values, JMX auth
credentials are enabled (and generated if necessary) and SSL is enabled.
If the user has not specified an SSL keystore/pass an SSL self-signed
certificate is generated, and either way the certificate is used for securing JMX
as well as HTTPS/WSS.