Proof-of-concept
Subprocess report generation so that if report requires too much memory,
the subprocess can be killed rather than allowing the whole container to
be taken down. This can be done by the parent process setting a smaller
maximum heap size for the subprocess (not yet configurable). The
subprocess is not able to connect back to container-jfr itself
currently.