Subprocess report generation (#286)
* Proof-of-concept

Subprocess report generation so that if report requires too much memory,
the subprocess can be killed rather than allowing the whole container to
be taken down. This can be done by the parent process setting a smaller
maximum heap size for the subprocess (not yet configurable). The
subprocess is not able to connect back to container-jfr itself
currently.

* Remove redundant parent process work performed by subprocess

* Limit maximum subprocess generation time

* Use Logger rather than raw System.out/err

* Spotless format

* Subprocess doesn't need to listen for incoming JMX connections

* fixup! Use Logger rather than raw System.out/err

* Better handling of subprocess timeout

* Update TODOS

* Write transformer set to file used for writing report

* Apply spotless formatting

* Refactor to use Builder with configurable env

* Pass optional credentials via env vars

* Simplify jvm arg list creation

* Use EpsilonGC for subprocess

* Add TODO

* Refactor cleanup

* Process archived recording reports in subprocess

* Fix tests

* Get truststore values from env, not filesystem

* Make /truststore dir configurable

* Use existing FileSystem local var

* Return Future and clean up clones

Return Future to allow better async handling of long-running subprocess report generation. Close recording streams when subprocess successfully generates reports to avoid dangling clone recordings, and if subprocess exits due to OOM, clean up dangling clones from parent process

* Refactor, add nonnull assertions, add tests

* Set subprocess OOM score adjustment if supported

* Allow null targetIds

This may be used in cases where a ConnectionDescriptor is used to represent contextual Credentials from an HTTP request for a non-target-oriented action or resource

* Address spotbugs violations

* Report target connection failure more accurately

* Remove subprocess max heap size

* Log elapsed subprocess run time

* Revert "Remove subprocess max heap size"

This reverts commit fb8252ac0355292d0c966d311313d743e06b230d.

If the subprocess does not exit due to exceeding the specified heap size
then the exit status will not be set as expected, and the parent process
will not know why the subprocess failed

* Ensure subprocess is terminated

* Add env var to configure subprocess max heap

* Document possible error message response on report generation

* Document CONTAINER_JFR_REPORT_GENERATION_MAX_HEAP

* Set subprocess timeout equal to HTTP timeout

* Copy recording to file temporarily

Open target connection and copy recording stream to file, then close target connection before proceeding with report generation. This holds the target connection open for the least time possible

* Improve archived report generation failure handling

Write error message to cached file and serve this to clients. This brings the behaviour in line with the failure handling for active recordings.