Remove io.kubernetes dependency
Remove io.kubernetes dependency and instead use already-present (transitive) io.fabric8.kubernetes. This also simplifies the type hierarchy between KubeApi and OpenShift platform classes, allowing for the removal of the OpenShiftPlatformClient and the substitution of the KubeApiPlatformClient, which is now a generified version of the OpenShiftPlatformClient implementation but using a KubernetesClient instance.