fix(archive): sanitize target alias when writing file
The target alias is currently included as-is in the destination filename when writing a JFR file to disk. This can be problematic since the target alias can contain characters like '/', for example `/deployments/quarkus-run.jar` in the quarkus-test sample application.  The solution presented here is to URL encode the alias before including it in the destination filename, so the alias is transformed into ex. `%2Fdeployments%2Fquarkus-run.jar`.