Initial rough cut
Adds API endpoints for pluggable discovery.

Also introduces DiscoveryStorage which is either a stand-in for a
database ORM or an abstraction around a later database connection to
come.

BuiltInDiscovery is used as a bridge between existing PlatformClient
implementations and platform-specific discovery behaviours and the
DiscoveryStorage component. This allows existing PlatformClients to
continue performing discovery within the main Cryostat process, but
publish those results into the DiscoveryStorage when changes occur. API
endpoints that query for a list of targets or the discovery tree read
from DiscoveryStorage so that these calls do not cascade out into
further network calls (ex. for KubeApiPlatformClient).