DiscoveryStorage implements PlatformClient
The new DiscoveryStorage replaces the previous platform detection
mechanism and provided PlatformClient, and the MergingPlatformClient.
The CustomTargetPlatformClient and the platform-detected appropriate
PlatformClient implementation are injected into a `Set<PlatformClient>`
which the BuiltInDiscovery consumes and uses to bridge updates into the
DiscoveryStorage. Since DiscoveryStorage itself implements the
PlatformClient interface, but is able to return results from storage
rather than re-querying the base platform, it is used as a replacement
injection in various existing places that injected the PlatformClient
instance previously (which would have always received the
MergingPlatformClient instance).