feat(discovery): pluggable discovery services API (#1026)
* Initial rough cut

Adds API endpoints for pluggable discovery.

Also introduces DiscoveryStorage which is either a stand-in for a
database ORM or an abstraction around a later database connection to
come.

BuiltInDiscovery is used as a bridge between existing PlatformClient
implementations and platform-specific discovery behaviours and the
DiscoveryStorage component. This allows existing PlatformClients to
continue performing discovery within the main Cryostat process, but
publish those results into the DiscoveryStorage when changes occur. API
endpoints that query for a list of targets or the discovery tree read
from DiscoveryStorage so that these calls do not cascade out into
further network calls (ex. for KubeApiPlatformClient).

* DiscoveryStorage implements PlatformClient

The new DiscoveryStorage replaces the previous platform detection
mechanism and provided PlatformClient, and the MergingPlatformClient.
The CustomTargetPlatformClient and the platform-detected appropriate
PlatformClient implementation are injected into a `Set<PlatformClient>`
which the BuiltInDiscovery consumes and uses to bridge updates into the
DiscoveryStorage. Since DiscoveryStorage itself implements the
PlatformClient interface, but is able to return results from storage
rather than re-querying the base platform, it is used as a replacement
injection in various existing places that injected the PlatformClient
instance previously (which would have always received the
MergingPlatformClient instance).

* fixup! Initial rough cut

* fixup! fixup! Initial rough cut

* Refactor and cleanup

* add discovery deregistration endpoint

* builtin discovery can be disabled

* add comment

* use UUID string for plugin registration ID, not numeric

* include placeholder auth token in registration response

* handle null children set as empty set

* internally-added callback URIs may be null

non-nullness is enforced on API endpoint. this allows internal cases like BuiltInDiscovery to register themselves through the discovery mechanism as an internal API client, without having to expose some meaningless callback URL since the Cryostat process is the "plugin" process

* experimental fixes

* don't send previous state back to plugin

* fix typos

* implement gson adapter for parsing subtrees provided by discovery plugins

* add quarkus-test sample using pluggable discovery

* builtin discovery completes startup promise

* discovery storage is also a verticle

* user basic auth

* fix image tag

* emit warnings on unexpected object members

clears spotbugs warning for missing 'default' in switch

* fixup webPort for quarkus-test-plugin