builtin discovery revamp
on startup, check if a realm is registered for each built-in discovery mechanism (platform client). register a plugin realm if not, then attach an update listener to the platform client to update that realm. delegate to the platform client to perform on-start load of current known state to replace existing persistent storage contents. Custom Targets overrides this and does not replace existing storage contents, so that the discovery database is the backing storage for custom targets