feat(discovery): implement JWT flow
On discovery plugin registration, server generates a JWT and supplies it to the plugin. This JWT must be supplied by the plugin in place of Authorization header for discovery updates and deregistration. Server's existing periodic ping task to check plugin reachability should prompt plugin to attempt to re-register itself, but this time it has an existing valid token to supply. If the server gets a registration request that includes a refreshed token is generated and supplied back into the same registration request flow.