feat(discovery): JWT flow (#1059)
* chore(jwt): rename things to prep for introducing another JWT flow

* feat(discovery): implement JWT flow

On discovery plugin registration, server generates a JWT and supplies it to the plugin. This JWT must be supplied by the plugin in place of Authorization header for discovery updates and deregistration. Server's existing periodic ping task to check plugin reachability should prompt plugin to attempt to re-register itself, but this time it has an existing valid token to supply. If the server gets a registration request that includes a refreshed token is generated and supplied back into the same registration request flow.

* refactoring

* remove outdated document

* move docs into new directory

* document discovery plugin API endpoints

* re-registration must match original realm and callback

* tokens expire after 2*ping_period

* fix unit tests

* minor fixes, updated itests

* fix spotbugs warnings

* fixup! fix spotbugs warnings

* throw 401 is token is expected and not provided

* update plugin ping logic

startup ping is a GET, later pruning pings are POST. plugins should simply respond with a 2xx status to a GET and sit idle, but after a POST they should respond with a 2xx and then call back to the Cryostat registration endpoint to refresh their access token

* use later quarkus-test that implements JWT flow

* fixup! update plugin ping logic

* fixup! use later quarkus-test that implements JWT flow

* document POST/GET ping difference

* time-expired tokens may still deregister plugins

* document quarkus-test-plugin 10010 port