fix(TargetConnectionManager): limit concurrent connections
Replace previous cache max size behaviour with max concurrent
connections behaviour. This is implemented using a semaphore. When a new
connection is opened a permit is first drained from the semaphore. When
the connection closes, whether explicitly, due to connection exception,
or due to cache TTL expiry, a permit is released.